requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
mongoengine==0.29.1
pymongo==4.15.4
pydantic==2.12.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
cryptography==46.0.3
python-multipart==0.0.20
dnspython==2.8.0
requests==2.31.0
